+++ lib/src/flexible_table_controller.dart
@@ -1,5 +1,6 @@
-import 'package:flexible_scrollable_table_view/src/scrollable/scroll_synchronization_mixin.dart';
+import 'package:flexible_scrollable_table_view/src/scrollable/synchronized_scroll_controller.dart';
 import 'package:flutter/foundation.dart';
+import 'package:flutter/widgets.dart';
 
 import 'selectable/selectable_row_mixin.dart';
 import 'sortable/sortable_column_mixin.dart';
@@ -14,9 +15,18 @@ abstract class FlexibleTableValueListenable<T> extends ValueListenable<T> {
 ///表控制器
 ///[T] 表数据实体
 class FlexibleTableController<T> extends ChangeNotifier
-    with SortableColumnMixin<T>, SelectableRowMixin<T>, ScrollSynchronizationMixin
+    with SortableColumnMixin<T>, SelectableRowMixin<T>
     implements FlexibleTableValueListenable<List<T>> {
   FlexibleTableController() : super();
+
+  ///用于横向滚动区域的同步滚动控制器
+  final ScrollController horizontalScrollController = SynchronizedScrollController();
+
+  @override
+  void dispose() {
+    horizontalScrollController.dispose();
+    super.dispose();
+  }
 
   ///原始数据
   List<T> _rawValue = <T>[];