+++ lib/src/widgets/flexible_table_header.dart
@@ -5,7 +5,6 @@ import 'package:flexible_scrollable_table_view/src/decoration/flexible_table_dec
 import 'package:flexible_scrollable_table_view/src/flexible_column.dart';
 import 'package:flexible_scrollable_table_view/src/flexible_table_configurations.dart';
 import 'package:flexible_scrollable_table_view/src/flexible_table_controller.dart';
-import 'package:flexible_scrollable_table_view/src/scrollable/horizontal_scroll_controller_builder.dart';
 import 'package:flexible_scrollable_table_view/src/widgets/flexible_table_header_cell.dart';
 import 'package:flutter/widgets.dart';
 
@@ -155,20 +154,17 @@ class ScrollableTableHeaderRow<T> extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return HorizontalScrollControllerBuilder(
-      arguments.controller,
-      builder: (context, scrollController) => ListView.builder(
-        controller: scrollController,
-        itemCount: arguments.scrollableColumnList.length,
-        scrollDirection: Axis.horizontal,
-        primary: false,
-        padding: EdgeInsets.zero,
-        physics: physics,
-        itemBuilder: (context, index) => FlexibleTableHeaderCell<T>(
-          arguments,
-          decorations: decorations,
-          column: arguments.scrollableColumnList[index],
-        ),
+    return ListView.builder(
+      controller: arguments.controller.horizontalScrollController,
+      itemCount: arguments.scrollableColumnList.length,
+      scrollDirection: Axis.horizontal,
+      primary: false,
+      padding: EdgeInsets.zero,
+      physics: physics,
+      itemBuilder: (context, index) => FlexibleTableHeaderCell<T>(
+        arguments,
+        decorations: decorations,
+        column: arguments.scrollableColumnList[index],
       ),
     );
   }